Move inner classes to outer
